feat: cut rav-appearance, so a theme reaches every surface
Ramp sat in the same file as `use tiny_skia`, under a doc comment saying the
renderer was meant to survive being made no_std. That was untrue as written, and
it is the second time in this restructure that a comment asserting a boundary
was false at the moment it was read.

Ink, Colour and Ramp now live in a crate that depends only on rav-core and names
no rasteriser, so the same theme dresses a glyph grid, a kitty image, a window
and an LED matrix. That is what makes an ESP32 with a screen look like rav
rather than merely work.

The layer hands back an Ink as readily as a Colour, because a terminal is the
one surface that cannot accept a resolved colour: the terminal and mono themes
exist to defer to the user's palette, and resolving early is what would break
them.

rav-core's math shim is public now. The sibling needed the same operations, and
two shims would be two chances to disagree - which is the failure this whole
split exists to make impossible.

It needs an allocator where the core does not, because Ramp owns its stops. That
is honest rather than ideal: an allocator-free target wants a Ramp that borrows
from a const, which is a change worth making against real hardware rather than
before it.
