+++ src/render/mod.rs
@@ -8,12 +8,13 @@
 //! allocator. What stays here is the part that cannot: the colour model, and the
 //! rasteriser that owns a `tiny-skia` pixmap.
 
+pub use rav_appearance::{ink, ramp};
 pub use rav_core::{capability, geometry};
 
-pub mod ink;
 pub mod raster;
 
 pub use capability::{Capabilities, Requirements, Shortfall};
 pub use geometry::{Anchor, BarLayout, Column, Rectangle, Screen};
 pub use ink::{Colour, Ink};
-pub use raster::{Band, Canvas, CapStyle, Ramp, Scene, Stripe, Style, StyleId};
+pub use ramp::{Ramp, Stripe};
+pub use raster::{Band, Canvas, CapStyle, Scene, Style, StyleId};
